Update README.md
Software of image renderer. Works on the Ray Tracing principle. A graphics set that includes geometric definition, camera, and various types of lighting. Developed with Java and was built on an agile software development approach that includes learning a language on the go, test-driven development (TDD), and program flexibility throughout its writing (refactoring).